% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwt_wst.R
\name{core}
\alias{core}
\title{Extracts the center of a vector}
\usage{
core(x, n)
}
\description{
This function is an internal function which extracts from a vector \code{x},
the center of the vector of length \code{n}. It emulates the Matlab(R) function \code{wkeep}.
This function is used by the cwt_wst function when the HAAR wavelet is selected.
}
