% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow2scales.R
\name{pow2scales}
\alias{pow2scales}
\title{Power 2 scales}
\usage{
pow2scales(scales)
}
\arguments{
\item{scales}{A vector of three elements with the minimum scale, the maximum scale and
the number ofsuboctaves per octave.}
}
\value{
A vector with all the scales.
}
\description{
This function constructs power 2 scales (i.e. using a base 2 logarithmic
scale) from a vector of three elements with the minimum scale, the maximum scale and
the number of suboctaves per octave, following Torrence and Compo 1998.
}
\section{References}{


C. Torrence, G. P. Compo. A practical guide to wavelet analysis. B. Am. Meteorol. Soc.
79 (1998), 61–78.
}

\examples{
scales <- pow2scales(c(2,128,8))

}
