% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_bases.R
\name{wave_bases}
\alias{wave_bases}
\title{Daughter wavelets}
\usage{
wave_bases(wname = c("MORLET", "PAUL", "DOG"), k, scale, wparam)
}
\arguments{
\item{wname}{A string, equal to "MORLET", "PAUL" or "DOG".}

\item{k}{A vector with the Fourier frequencies at which to calculate the wavelet.}

\item{scales}{A vector with the wavelet scales.}

\item{wparam}{The corresponding nondimensional parameter for the wavelet function.}
}
\value{
A matrix of size length(scale) x length(k), \code{daughter},
containing the fourier transform of the daughter wavelets for the mother "wname",
for the given scales (one row for each scale).
}
\description{
This function computes the daughter wavelets for three different families: "MORLET", "DOG" and "PAUL"
(Derivatives of the Gaussian). It is a translation from the Matlab(R) function published by Torrence
and Compo following (Torrence & Compo, 1998).

This is an internal function used for the computation of the wavelet continuous
transform with the \code{cwt_wst} function
}
\section{References}{

C. Torrence, G. P. Compo. A practical guide to wavelet analysis. B. Am. Meteorol. Soc.
79 (1998), 61–78.
}

\examples{
k <- (1:2000)*2/pi/100
scales <- c(1,2)
daughter <- wave_bases(wname = "MORLET", k = k, scale = scales)
plot(k, daughter[2,], type = "l", xlab = "", ylab = "", main = "Morlet daughter wavelets")
lines(k,daughter[1,], col = "red")
legend("topright", legend = c("s = 1", "s = 2"), lty = 1, col = c("red", "black"))


}
